+++ circumflex-orm/src/main/scala/validation.scala
@@ -2,8 +2,11 @@ package ru.circumflex.orm
 
 import ru.circumflex.core._
 
-class ValidationException(val errors: MsgGroup)
-    extends ORMException("Record validation failed.")
+class ValidationException(val errors: Seq[Msg])
+    extends ORMException("Record validation failed.") {
+  def this(msg: Msg) = this(List(msg))
+  def this(msg: Msg, msgs: Msg*) = this(List(msg) ++ msgs.toSeq)
+}
 
 class RecordValidator[PK, R <: Record[PK, R]] {